requirements.txt
aiosqlite==0.19.0
bidict==0.21.2
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
eventlet==0.36.1
Flask==2.2.2
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
itsdangerous==2.0.1
Jinja2==3.1.3
MarkupSafe==2.1.3
numexpr==2.8.4
numpy==1.26.4
packaging==24.0
pandas==2.0.3
pycparser==2.21
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2023.3.post1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.8
win-inet-pton==1.1.0
wsproto==1.2.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.4.post2